# api/docs/swagger.yaml
basePath: /
definitions:
...
  controllers.RegisterInput:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - name
    - password
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Backend multi-tenant for electronics shops
  title: MultiShop API
  version: "1.0"
paths:
...
  /api/users:
    get:
      description: Allows a SuperAdmin to retrieve all users belonging to their shop.
        Only id, email and role are returned.
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - SuperAdmin only
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all users of the current shop
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Allows a SuperAdmin to create a new Admin user in their own shop.
        Role is automatically set to Admin.
      parameters:
      - description: Admin creation payload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controllers.RegisterInput'
      produces:
      - application/json
      responses:
        "201":
          description: Admin successfully created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input or creation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - SuperAdmin only
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new Admin user
      tags:
      - Users
  /api/users/{id}:
    delete:
      description: Allows a SuperAdmin to delete a user belonging to their shop.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: User not found
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - SuperAdmin only
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a user
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Allows a SuperAdmin to update the role of a user belonging to their
        shop. Only Admin or SuperAdmin roles are allowed.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Role update payload (Admin or SuperAdmin)
        in: body
        name: input
        required: true
        schema:
          properties:
            role:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Role updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid role or user not found
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - SuperAdmin only
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update user role
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"